==> Helina Project V2/Jory WT 10 Chemicals/3.1_R-ODAF_DEG_Median BMD List Extraction.R <==
####Libraries####
library(tidyverse)
source("BMDExpressFunctions.R")

####Metadata Import####
metadata <- read.csv("RNAseqData/metadata.csv")
chemnames <- unique(metadata$chemical)
lowestdoses <- unique(metadata[, c("chemical", "dose")]) %>%
  group_by(chemical) %>%
  filter(dose > 0) %>%
  summarise_all(min)

####GO Term BMD Extraction####
#Import
go_raw_data <- raw_importer(file_path = "BMDExpressData_R-ODAF/GO_TERM/R_ODAF_DEGs/", 
                            file_type = ".txt", 
                            start_phrase = "GO/Pathway/Gene Set/Gene ID") %>%
  lapply(cleanupcolumns_goterm) %>%
  setNames(chemnames)

#Filtering
go_raw_data_filtered <- list()
for (i in 1:length(go_raw_data)) {
  go_raw_data_filtered[[chemnames[i]]] <-
    goterm_filtering(x = go_raw_data[[i]]) %>%
    mutate(logBMD = log10(BMD.Median)) %>%
    arrange(logBMD)
}

#Filtering 2: Electric Boogaloo
go_BMD_list <- list()
for (i in chemnames) {
  if (nrow(go_raw_data_filtered[[i]] > 0)) {
    go_BMD_list[[i]] <- BMD_list_extraction(go_raw_data_filtered[[i]])
  }
}

go_BMD_list_logtransformed <- list()
for(i in chemnames){
  go_BMD_list_logtransformed[[i]] <- lapply(go_BMD_list[[i]], log10)
}

####Reactome BMD Extraction####
#Import
reactome_raw_data <- raw_importer(file_path = "BMDExpressData_R-ODAF/REACTOME/R-ODAF_DEGs/", 
                                     file_type = ".txt", 
                                     start_phrase = "GO/Pathway/Gene Set/Gene ID") %>%
  lapply(cleanupcolumns_reactome) %>%
  setNames(chemnames)

#Filtering
reactome_raw_data_filtered <- list()
for (i in 1:length(reactome_raw_data)) {
  reactome_raw_data_filtered[[chemnames[i]]] <-
    reactome_filtering(x = reactome_raw_data[[i]]) %>%
    mutate(logBMD = log10(BMD.Median)) %>%
    arrange(logBMD)
}

#Filtering 2: Electric Boogaloo
reactome_BMD_list <- list()
for (i in chemnames) {
  if (nrow(reactome_raw_data_filtered[[i]] > 0)) {
    reactome_BMD_list[[i]] <-
      BMD_list_extraction(reactome_raw_data_filtered[[i]])
  }

}

reactome_BMD_list_logtransformed <- list()
for(i in chemnames){
  reactome_BMD_list_logtransformed[[i]] <- lapply(reactome_BMD_list[[i]], log10)
}

####Full BMD list####

BMD_raw_data <- list()
BMD_filenames <- list.files("BMDExpressData_R-ODAF/BMD/R-ODAF_DEGS", pattern = ".txt")

# for(i in 1:length(chemnames)){
#   BMD_raw_data[[chemnames[i]]] <- read.table(paste0("BMDExpressData/BMD/",BMD_filenames[i]), header = TRUE, sep = "\t") %>%
#     cleanupcolumns()
# }

BMD_raw_data <- raw_importer(file_path = "BMDExpressData_R-ODAF/BMD/R-ODAF_DEGS/", 
                             file_type = ".txt", 
                             start_phrase = "Probe Id") %>%
  lapply(cleanupcolumns) %>%
  setNames(chemnames)


# Filter
all_BMD_list_logtransformed <- list()
for (i in chemnames) {
  all_BMD_list_logtransformed[[i]] <-
    BMDfiltering(x = BMD_raw_data[[i]],
                 lowdose = lowestdoses$dose[lowestdoses$chemical == i],
                 highdose = 3) %>%
    mutate(logBMD = log10(BMD)) %>%
    select(logBMD) %>%
    pull() %>%
    list()
}

####Export Data#### 
saveRDS(go_BMD_list_logtransformed, "go_BMD_list_logtransformed_R-ODAF_DEGs.RDS")
saveRDS(reactome_BMD_list_logtransformed, "reactome_BMD_list_logtransformed_R-ODAF_DEGs.RDS")
saveRDS(all_BMD_list_logtransformed, "all_BMD_list_logtransformed_R-ODAF_DEGs.RDS")
